feat: move under `refactor.internal.graph_access`
