+++ refactor/internal/graph_access.py
